Make use of extracted `useSearchBarSubmit`, consolidate logic.
