+++ graylog2-web-interface/src/views/components/DashboardSearchBarForm.tsx
@@ -15,7 +15,7 @@
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
 import * as React from 'react';
-import { useCallback, useContext, useState } from 'react';
+import { useCallback, useContext } from 'react';
 import PropTypes from 'prop-types';
 import { Form, Formik } from 'formik';
 import isFunction from 'lodash/isFunction';
@@ -29,7 +29,7 @@ import { isNoTimeRangeOverride } from 'views/typeGuards/timeRange';
 import usePluginEntities from 'hooks/usePluginEntities';
 import useUserDateTime from 'hooks/useUserDateTime';
 import useHandlerContext from 'views/components/useHandlerContext';
-import { normalizeFromSearchBarForBackend } from 'views/logic/queries/NormalizeTimeRange';
+import useSearchBarSubmit from 'views/components/searchbar/useSearchBarSubmit';
 
 import { onInitializingTimerange } from './TimerangeForForm';
 
@@ -49,25 +49,18 @@ type Props = {
 const _isFunction = (children: Props['children']): children is (props: FormikProps<DashboardFormValues>) => React.ReactElement => isFunction(children);
 
 const DashboardSearchForm = ({ initialValues, limitDuration, onSubmit, validateQueryString, children }: Props) => {
-  const { formatTime, userTimezone } = useUserDateTime();
+  const { formatTime } = useUserDateTime();
   const { setFieldWarning } = useContext(FormWarningsContext);
-  const [enableReinitialize, setEnableReinitialize] = useState(true);
   const pluggableSearchBarControls = usePluginEntities('views.components.searchBar');
 
-  const _onSubmit = useCallback(({ timerange, ...rest }: DashboardFormValues) => {
-    setEnableReinitialize(false);
-
-    return onSubmit({
-      timerange: isNoTimeRangeOverride(timerange) ? undefined : normalizeFromSearchBarForBackend(timerange, userTimezone),
-      ...rest,
-    }).then(() => setEnableReinitialize(true));
-  }, [onSubmit, userTimezone]);
   const { timerange, ...rest } = initialValues;
   const initialTimeRange = timerange && !isNoTimeRangeOverride(timerange) ? onInitializingTimerange(timerange, formatTime) : {} as TimeRange;
   const _initialValues = {
     timerange: initialTimeRange,
     ...rest,
   };
+
+  const { enableReinitialize, onSubmit: _onSubmit } = useSearchBarSubmit(_initialValues, onSubmit);
 
   const handlerContext = useHandlerContext();
   const _validate = useCallback((values: DashboardFormValues) => validate(values, limitDuration, setFieldWarning, validateQueryString, pluggableSearchBarControls, formatTime, handlerContext),
