+++ graylog2-web-interface/src/views/components/WidgetQueryControls.tsx
@@ -60,13 +60,15 @@ import { updateWidget } from 'views/logic/slices/widgetActions';
 import { execute, setGlobalOverrideQuery, setGlobalOverrideTimerange } from 'views/logic/slices/searchExecutionSlice';
 import useAppDispatch from 'stores/useAppDispatch';
 import useHandlerContext from 'views/components/useHandlerContext';
+import { isNoTimeRangeOverride } from 'views/typeGuards/timeRange';
+import { normalizeFromSearchBarForBackend } from 'views/logic/queries/NormalizeTimeRange';
 
 import TimeRangeOverrideInfo from './searchbar/WidgetTimeRangeOverride';
 import TimeRangeFilter from './searchbar/time-range-filter';
 import StreamsFilter from './searchbar/StreamsFilter';
 import SearchButton from './searchbar/SearchButton';
 import QueryInput from './searchbar/queryinput/AsyncQueryInput';
-import SearchBarForm, { normalizeSearchBarFormValues } from './searchbar/SearchBarForm';
+import SearchBarForm from './searchbar/SearchBarForm';
 import WidgetQueryOverride from './WidgetQueryOverride';
 import PluggableSearchBarControls from './searchbar/PluggableSearchBarControls';
 
@@ -115,10 +117,13 @@ const useBindApplySearchControlsChanges = (formRef) => {
   useEffect(() => {
     bindApplySearchControlsChanges((newWidget: Widget) => {
       if (formRef.current) {
-        const { dirty, values, isValid } = formRef.current;
+        const { dirty, values: { timerange, ...rest }, isValid } = formRef.current;
 
         if (dirty && isValid) {
-          const normalizedFormValues = normalizeSearchBarFormValues(values, userTimezone);
+          const normalizedFormValues = {
+            timerange: isNoTimeRangeOverride(timerange) ? undefined : normalizeFromSearchBarForBackend(timerange, userTimezone),
+            ...rest,
+          };
 
           return updateWidgetSearchControls(newWidget, normalizedFormValues);
         }
