+++ graylog2-web-interface/src/views/components/searchbar/SearchBarForm.tsx
@@ -15,15 +15,13 @@
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
 import * as React from 'react';
-import { useCallback, useContext, useMemo, useState } from 'react';
+import { useCallback, useContext, useMemo } from 'react';
 import PropTypes from 'prop-types';
 import type { FormikProps } from 'formik';
 import { Form, Formik } from 'formik';
 import isFunction from 'lodash/isFunction';
 
-import { SearchQueryStrings } from '@graylog/server-api';
 import { onInitializingTimerange } from 'views/components/TimerangeForForm';
-import { normalizeFromSearchBarForBackend } from 'views/logic/queries/NormalizeTimeRange';
 import type { SearchBarFormValues } from 'views/Constants';
 import FormWarningsContext from 'contexts/FormWarningsContext';
 import type { QueryValidationState } from 'views/components/searchbar/queryvalidation/types';
@@ -31,6 +29,7 @@ import validate from 'views/components/searchbar/validate';
 import usePluginEntities from 'hooks/usePluginEntities';
 import useUserDateTime from 'hooks/useUserDateTime';
 import useHandlerContext from 'views/components/useHandlerContext';
+import useSearchBarSubmit from 'views/components/searchbar/useSearchBarSubmit';
 
 type FormRenderer = (props: FormikProps<SearchBarFormValues>) => React.ReactNode;
 type Props = {
@@ -45,26 +44,8 @@ type Props = {
 
 const _isFunction = (children: Props['children']): children is FormRenderer => isFunction(children);
 
-export const normalizeSearchBarFormValues = ({ timerange, ...rest }: SearchBarFormValues, userTimezone: string) => ({ timerange: normalizeFromSearchBarForBackend(timerange, userTimezone), ...rest });
-
-const executeWithQueryStringRecording = async <R, >(isDirty: boolean, query: string, callback: () => R) => {
-  const trimmedQuery = query.trim();
-
-  try {
-    if (isDirty && !!trimmedQuery) {
-      await SearchQueryStrings.queryStringUsed({ query_string: trimmedQuery });
-    }
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.error('Unable to record last used query string: ', error);
-  }
-
-  return callback();
-};
-
 const SearchBarForm = ({ initialValues, limitDuration, onSubmit, children, validateOnMount, formRef, validateQueryString }: Props) => {
-  const [enableReinitialize, setEnableReinitialize] = useState(true);
-  const { formatTime, userTimezone } = useUserDateTime();
+  const { formatTime } = useUserDateTime();
   const pluggableSearchBarControls = usePluginEntities('views.components.searchBar');
   const { setFieldWarning } = useContext(FormWarningsContext);
   const _initialValues = useMemo(() => {
@@ -76,16 +57,7 @@ const SearchBarForm = ({ initialValues, limitDuration, onSubmit, children, valid
     });
   }, [formatTime, initialValues]);
 
-  const _onSubmit = useCallback((values: SearchBarFormValues) => {
-    setEnableReinitialize(false);
-    const queryString = values?.queryString;
-
-    return executeWithQueryStringRecording(
-      queryString !== _initialValues?.queryString,
-      queryString,
-      () => onSubmit(normalizeSearchBarFormValues(values, userTimezone)).finally(() => setEnableReinitialize(true)),
-    );
-  }, [_initialValues.queryString, onSubmit, userTimezone]);
+  const { enableReinitialize, onSubmit: _onSubmit } = useSearchBarSubmit(_initialValues, onSubmit);
 
   const handlerContext = useHandlerContext();
   const _validate = useCallback((values: SearchBarFormValues) => validate(values, limitDuration, setFieldWarning, validateQueryString, pluggableSearchBarControls, formatTime, handlerContext),
